calculate average using speedtest-go

+++ fastcom/fastcom.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 )
 
-func TestSpeed() (int, int, error) {
+func TestSpeed() (float64, float64, error) {
 	fmt.Println("Hello from fast.com gauge")
 	return 1, 1, nil
 }
